requirements.txt
selenium
pillow
pyzbar
qrcode